♻️ Extract `useViewsCount` into a custom hook
