+++ src/app/blog/components/Post/ViewsCount/index.tsx
@@ -1,28 +1,11 @@
 'use client'
 
-import useSWR from 'swr'
+import useViewsCount from 'src/utils/hooks/useViewsCount'
 
 type Props = { slug: string; views?: string }
 
-async function fetcher<JSON>(
-  input: RequestInfo,
-  init?: RequestInit,
-): Promise<JSON> {
-  const response = await fetch(input, init)
-
-  return response.json()
-}
-
 const ViewsCount = ({ slug, views }: Props) => {
-  const { data } = useSWR<{ views: Record<string, string> }>(
-    '/api/views',
-    fetcher,
-    views
-      ? {
-          fallbackData: { views: { [slug]: views } },
-        }
-      : undefined,
-  )
+  const { data } = useViewsCount(slug, views)
 
   return <>{data?.views?.[slug]}</>
 }
